requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
jinja2==3.1.5
psycopg2-binary==2.9.10
python-multipart==0.0.20
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7
alembic==1.16.2
python-dotenv==1.1.1
pydantic_settings==2.10.1